[mathcore] Move RANLUX++ implementation to directory
Split out addition and subtraction related functions to helpers.h,
they will be re-used in future changes. Also add header guards and
make the functions in mulmod.h static to avoid redefinition errors
should the file be included into multiple TUs.

(cherry picked from commit 42a224dee93b37a67323233ca51e162ac52f6bab)
