+++ math/mathcore/src/ranluxpp/mulmod.h
@@ -9,52 +9,19 @@
  * For the list of contributors see $ROOTSYS/README/CREDITS.             *
  *************************************************************************/
 
+#ifndef RANLUXPP_MULMOD_H
+#define RANLUXPP_MULMOD_H
+
+#include "helpers.h"
+
 #include <cstdint>
-
-/// Compute `a + b` and set `overflow` accordingly.
-static inline uint64_t add_overflow(uint64_t a, uint64_t b, unsigned &overflow)
-{
-   uint64_t add = a + b;
-   overflow = (add < a);
-   return add;
-}
-
-/// Compute `a + b` and increment `carry` if there was an overflow
-static inline uint64_t add_carry(uint64_t a, uint64_t b, unsigned &carry)
-{
-   unsigned overflow;
-   uint64_t add = add_overflow(a, b, overflow);
-   // Do NOT branch on overflow to avoid jumping code, just add 0 if there was
-   // no overflow.
-   carry += overflow;
-   return add;
-}
-
-/// Compute `a - b` and set `overflow` accordingly
-static inline uint64_t sub_overflow(uint64_t a, uint64_t b, unsigned &overflow)
-{
-   uint64_t sub = a - b;
-   overflow = (sub > a);
-   return sub;
-}
-
-/// Compute `a - b` and increment `carry` if there was an overflow
-static inline uint64_t sub_carry(uint64_t a, uint64_t b, unsigned &carry)
-{
-   unsigned overflow;
-   uint64_t sub = sub_overflow(a, b, overflow);
-   // Do NOT branch on overflow to avoid jumping code, just add 0 if there was
-   // no overflow.
-   carry += overflow;
-   return sub;
-}
 
 /// Multiply two 576 bit numbers, stored as 9 numbers of 64 bits each
 ///
 /// \param[in] in1 first factor as 9 numbers of 64 bits each
 /// \param[in] in2 second factor as 9 numbers of 64 bits each
 /// \param[out] out result with 18 numbers of 64 bits each
-void multiply9x9(const uint64_t *in1, const uint64_t *in2, uint64_t *out)
+static void multiply9x9(const uint64_t *in1, const uint64_t *in2, uint64_t *out)
 {
    uint64_t next = 0;
    unsigned nextCarry = 0;
@@ -169,7 +136,7 @@ void multiply9x9(const uint64_t *in1, const uint64_t *in2, uint64_t *out)
 ///
 /// Note that this function does *not* return the smallest value congruent to
 /// the modulus, it only guarantees a value smaller than \f$ 2^{576} \$!
-void mod_m(const uint64_t *mul, uint64_t *out)
+static void mod_m(const uint64_t *mul, uint64_t *out)
 {
    uint64_t r[9] = {0};
 
@@ -314,7 +281,7 @@ void mod_m(const uint64_t *mul, uint64_t *out)
 ///
 /// \param[in] in1 first factor with 9 numbers of 64 bits each
 /// \param[inout] inout second factor and also the output of the same size
-void mulmod(const uint64_t *in1, uint64_t *inout)
+static void mulmod(const uint64_t *in1, uint64_t *inout)
 {
    uint64_t mul[2 * 9] = {0};
    multiply9x9(in1, inout, mul);
@@ -328,7 +295,7 @@ void mulmod(const uint64_t *in1, uint64_t *inout)
 /// \param[in] n exponent
 ///
 /// The arguments base and res may point to the same location.
-void powermod(const uint64_t *base, uint64_t *res, uint64_t n)
+static void powermod(const uint64_t *base, uint64_t *res, uint64_t n)
 {
    uint64_t fac[9] = {0};
    fac[0] = base[0];
@@ -351,3 +318,5 @@ void powermod(const uint64_t *base, uint64_t *res, uint64_t n)
       mod_m(mul, fac);
    }
 }
+
+#endif
